Glitchy color changes
Initially set out on making a custom HTML element for facilitating
color changes. Realized quickly that that wasn't the right approach,
and started using configs to deal with color changes.

Color changes work, but are still very glitchy. The primary error
right now is that if you "turn on" the package twice, then turning
the package off does nothing. I'm not entirely sure as to what's
causing this issue, but one thing that might work to resolve it is
to start automatic codewall processing as opposed to manually turning
it on and off to refresh it.