Working Package
I've worked out most of the bugs with this package, and it should
be ready for basic use.

Main points:

Removed options to change highlight color / ruler length.
Not only are these options extraneous and outrageously expensive in
terms of their value-to-bug-creation-potential, but they aren't
main priorities for this package.

Added automatic re-checking whenever a change is made in an active text
editor. This means that users no longer need to turn codewall off and
on again whenever they make changes; changes are dynamically checked
while codewall is on.